more work on ges solver
